rename : hook 위치 변경
